+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - pyrotorque: New `ItemCommand` job
 - Basic tab completion
+- Optional `guessit_*` fields
 
 ### Changed
 - Warn if rTorrent doesn't support `string.contains_i` for fast